+++ resources/views/layouts/admin/aside.blade.php
@@ -4,10 +4,15 @@
         <div class="h-full px-3 pb-4 overflow-y-auto bg-white dark:bg-gray-900">
             <ul class="space-y-2">
 
-            <x-aside.dropdown label="Allgemein" svg="svg.office" >
+            <x-aside.dropdown label="Benutzer" svg="svg.user" >
                 <x-slot:links>
                     <x-aside.dropdownlink label="Benutzer" href="{{ route('admin.user.index') }}" />
                     <x-aside.dropdownlink label="Rollen" href="{{ route('admin.role.index') }}" />
+                </x-slot:links>
+            </x-aside.dropdown>
+
+            <x-aside.dropdown label="Kunden" svg="svg.office" >
+                <x-slot:links>
                     <x-aside.dropdownlink label="Kunden" href="{{ route('admin.customer.index') }}" />
                 </x-slot:links>
             </x-aside.dropdown>
@@ -15,6 +20,7 @@
             <x-aside.dropdown label="Auswahlmenüs" svg="svg.computer" >
                 <x-slot:links>
                     <x-aside.dropdownlink label="Betriebsysteme" href="{{ route('admin.operatingsystem.index') }}" />
+                    <x-aside.dropdownlink label="Mail Anbieter" href="{{ route('admin.mailboxprovider.index') }}" />
                 </x-slot:links>
             </x-aside.dropdown>