+++ routes/api.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\API\CustomerController;
 use App\Http\Controllers\API\RackCabinetController;
 use App\Http\Controllers\API\RackDeviceController;
 use App\Http\Controllers\API\RoomController;
+use App\Http\Controllers\Api\ServerController;
 use App\Http\Controllers\API\SiteController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -52,6 +53,17 @@ Route::middleware(['auth:sanctum'])->group(function () {
 
     // RackDevices
     Route::get('/rackdevices', [RackDeviceController::class, 'getAll']);
+
+
+
+
+
+    // Server
+    Route::get('servers', [ServerController::class, 'index']);
+    Route::get('servers/{server}', [ServerController::class, 'show']);
+    Route::post('servers', [ServerController::class, 'store']);
+    Route::put('servers/{server}', [ServerController::class, 'update']);
+    Route::delete('servers/{server}', [ServerController::class, 'delete']);
 
 
 });